authorized_networks routin refactored
