+++ src/xpk/commands/batch.py
@@ -15,7 +15,7 @@ limitations under the License.
 """
 
 from argparse import Namespace
-from ..utils import xpk_exit, xpk_print
+from ..utils.console import xpk_exit, xpk_print
 from .cluster import set_cluster_command
 from ..core.core import add_zone_and_project
 from ..core.job_template import create_job_template_instance
