+++ src/xpk/core/cluster_private.py
@@ -42,35 +42,6 @@ def authorize_private_cluster_access_if_necessary(args) -> int:
       )
       return 1
 
-  return_code, authorized_networks = get_authorized_networks_to_update(args)
-  if return_code != 0:
-    xpk_print('Error on getting authorized networks to update.')
-    return return_code
-
-  if authorized_networks is None:
-    xpk_print('Current machine is already in authorized networks.')
-    return 0
-
-  return_code, authorized_networks = add_current_machine_to_networks(
-      authorized_networks
-  )
-  if return_code != 0:
-    xpk_print(
-        "Adding current machine's IP address to the authorized networks failed!"
-    )
-    return return_code
-
-  cluster_authorized_networks_update_code = update_cluster_authorized_networks(
-      args, authorized_networks
-  )
-  if cluster_authorized_networks_update_code > 0:
-    xpk_print('Updating cluster authorized networks failed!')
-    return cluster_authorized_networks_update_code
-  xpk_print("Cluster's master authorized networks updated successfully.")
-  return 0
-
-
-def get_authorized_networks_to_update(args):
   new_authorized_networks_needed = args.authorized_networks is not None
 
   authorized_networks = (
@@ -79,17 +50,36 @@ def get_authorized_networks_to_update(args):
       else get_cluster_authorized_networks(args)
   )
 
-  if not new_authorized_networks_needed:
-    (
-        is_current_machine_in_any_network_return_code,
-        current_machine_already_authorized,
-    ) = is_current_machine_in_any_network(authorized_networks)
-    if is_current_machine_in_any_network_return_code > 0:
-      xpk_print("Error checking current machine's IP address.")
-      return is_current_machine_in_any_network_return_code, None
-    new_authorized_networks_needed = not current_machine_already_authorized
+  is_current_machine_in_network_return_code, is_current_machine_in_network = (
+      is_current_machine_in_any_network(authorized_networks)
+  )
+  if is_current_machine_in_network_return_code != 0:
+    xpk_print("Error on checking current machine's IP adrress.")
+    return is_current_machine_in_network_return_code
 
-  return 0, authorized_networks if new_authorized_networks_needed else None
+  if not is_current_machine_in_network:
+    add_current_machine_to_networks_return_code, authorized_networks = (
+        add_current_machine_to_networks(authorized_networks)
+    )
+    if add_current_machine_to_networks_return_code != 0:
+      xpk_print(
+          "Adding current machine's IP address to the authorized networks"
+          ' failed!'
+      )
+      return add_current_machine_to_networks_return_code
+
+  if new_authorized_networks_needed or not is_current_machine_in_network:
+    cluster_authorized_networks_update_code = (
+        update_cluster_authorized_networks(args, authorized_networks)
+    )
+    if cluster_authorized_networks_update_code != 0:
+      xpk_print('Updating cluster authorized networks failed!')
+      return cluster_authorized_networks_update_code
+    xpk_print("Cluster's master authorized networks updated successfully.")
+    return 0
+
+  xpk_print("Current machine's IP adrress is already authorized.")
+  return 0
 
 
 def is_cluster_private(args) -> bool:
